Cast pipeline to float32 in zuki_cute_ill_v50 load_pipe
The SDXL checkpoint contains float16 weights. Explicitly cast to
float32 when moving to CPU to avoid dtype mismatches during text
encoding preprocessing.

+++ zuki_cute_ill_v50/pytorch/src/model_utils.py
@@ -25,8 +25,7 @@ def load_pipe(variant):
     pipe = DiffusionPipeline.from_pretrained(variant, torch_dtype=torch.float32)
     modules = [pipe.text_encoder, pipe.unet, pipe.text_encoder_2, pipe.vae]
 
-    # Move the pipeline to CPU
-    pipe.to("cpu")
+    pipe.to("cpu", dtype=torch.float32)
 
     for module in modules:
         module.eval()
